chore(CollectionMenu): move modalWrapper css for conflicts
adds collectionTrigger css as well to resolve merge conflicts
